move format functions to utils
