+++ libs/feature/record/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/feature-record.module'
 export * from './lib/state'
 export * from './lib/links/link-classifier.service'
+export * from './lib/links/link-utils'
